feat: move fns out of test
