+++ src/utils/__tests__/cli.js
@@ -1,65 +1,5 @@
-const { getArgFromCLI, getArgsFromCLI } = require( '../cli' );
+const { getArg, getValueFromList, getValue, getArgsFromCLI } = require( '../cli' );
 const path = require( 'path' );
-
-const getValue = ( config, term ) => {
-
-	for (const key in config) {
-
-		if ( config.hasOwnProperty(key) ) {
-
-			if ( term === key ) {
-				return 'undefined' === typeof config[term] ? true : config[term];
-			}
-		}
-
-		if ( 'object' === typeof config[key] ) {
-			return getValueFromList( config[key], term );
-		}
-
-	}
-};
-
-const getValueFromList = ( list, term ) => {
-
-	if ( 0 === list.length ) {
-		return false;
-	}
-
-	const currItem = list[0];
-
-	if ( term === currItem ) {
-		return true;
-	}
-
-	if ( term === currItem[0] ) {
-		return currItem[1];
-	}
-
-	list.shift();
-
-	return getValueFromList( list, term );
-
-};
-
-const getArg = ( arg ) => {
-
-	const config = require( path.join( process.cwd(), 'css-audit.config.js' ) );
-
-	for ( const cliArg of getArgsFromCLI() ) {
-
-		const [ name, value ] = cliArg.split( '=' );
-
-		if ( name === arg ) {
-			return 'undefined' === typeof value ? true : value || null;
-		}
-
-	}
-
-	const term = arg.substr(2);
-
-	return getValue( config, term );
-
-};
 
 describe( 'Get args', () => {
